Trac #29169: tweak two max/critical angle error messages.
The phrase "this cone cannot be..." can be misinterpreted as making a
value judgment about the given cone, rather than a blanket statement
about the argument (self) to these two functions. This commit changes
it to "cone should not be..." to indicate that the user should not
pass in either a trivial cone or the ambient space.