Trac #30605: catch more coercion failures in cone._ambient_space_point().
The _ambient_space_point() function tries to coerce a point into the
ambient space with a series of try...except blocks that attempt less
and less desirable alternatives. The specific exception handled in the
"except" blocks is not the only one that can arise, however, leading
to some failures in cases like the following:

  sage: from sage.geometry.cone import _ambient_space_point
  sage: c = Cone([(1,0), (0,1)])
  sage: _ambient_space_point(c, vector(AA,[1/2,1/sqrt(3)]))
  ...
  ValueError: Cannot coerce non-integral Algebraic Real 1/2 to Integer

The unexpected `ValueError` is thrown before the expected `TypeError`,
so the function fails. This commit adds the `ValueError` above to the
list of expected exceptions, so that the function can proceed to try
more ambient spaces afterwards.